requirements.txt
# ── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Validation & Settings ────────────────────────────────────────────────────
pydantic[email]==2.10.3
pydantic-settings==2.7.0

# ── Database ─────────────────────────────────────────────────────────────────
supabase==2.10.0

# ── Cache / Queue ────────────────────────────────────────────────────────────
redis==5.2.1
arq==0.26.1

# ── Telegram Bot ─────────────────────────────────────────────────────────────
python-telegram-bot==21.9

# ── Email ────────────────────────────────────────────────────────────────────
aiosmtplib==3.0.2

# ── Cryptography ─────────────────────────────────────────────────────────────
bcrypt==4.2.1
cryptography==44.0.0
python-jose[cryptography]==3.3.0

# ── Templating ───────────────────────────────────────────────────────────────
jinja2==3.1.5

# ── HTTP Client ──────────────────────────────────────────────────────────────
# Note: supabase==2.10.0 requires httpx<0.28, so we pin to 0.27.x
httpx==0.27.2

# ── Security Headers ─────────────────────────────────────────────────────────
secure==0.3.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.2

# ── Linting / Typing ─────────────────────────────────────────────────────────
ruff==0.8.6
mypy==1.14.1
